refactor: move CSG tests into individual files
