+++ primap2/tests/csg/test_compose.py
@@ -1,7 +1,5 @@
 """Tests for csg/_compose.py"""
 
-from collections.abc import Sequence
-
 import numpy as np
 import pandas as pd
 import pytest
@@ -9,173 +7,7 @@ import xarray as xr
 
 import primap2.csg
 import primap2.csg._compose
-
-
-def get_single_ts(
-    *,
-    time: pd.DatetimeIndex | None = None,
-    data: np.ndarray | None = None,
-    dims: Sequence[str] | None = None,
-    coords: dict[str, str | Sequence[str]] | None = None,
-) -> xr.DataArray:
-    if time is None:
-        time = pd.date_range("1850-01-01", "2022-01-01", freq="YS")
-    if dims is None:
-        dims = []
-    if data is None:
-        data = np.linspace(0.0, 1.0, len(time))
-    if coords is None:
-        coords = {}
-    return xr.DataArray(
-        data,
-        dims=["time", *dims],
-        coords={"time": time, **coords},
-    )
-
-
-def test_substitution_strategy():
-    ts = get_single_ts(data=1.0)
-    ts[0] = np.nan
-    fill_ts = get_single_ts(data=2.0)
-
-    (
-        result_ts,
-        result_descriptions,
-    ) = primap2.csg.SubstitutionStrategy().fill(
-        ts=ts, fill_ts=fill_ts, fill_ts_repr="B"
-    )
-    assert result_ts[0] == 2.0
-    assert (result_ts[1:] == 1.0).all()
-    assert len(result_descriptions) == 1
-    assert result_descriptions[0].time == np.array(["1850"], dtype=np.datetime64)
-    assert (
-        result_descriptions[0].description
-        == "substituted with corresponding values from B"
-    )
-    assert "source" not in result_ts.coords.keys()
-
-
-def test_selector_match():
-    da = get_single_ts(coords={"source": "A", "category": "1.A"})
-
-    assert primap2.csg.StrategyDefinition.match(selector={"source": "A"}, fill_ts=da)
-    assert not primap2.csg.StrategyDefinition.match(
-        selector={"source": "B"}, fill_ts=da
-    )
-    assert primap2.csg.StrategyDefinition.match(
-        selector={"source": "A", "category": "1.A"}, fill_ts=da
-    )
-    assert primap2.csg.StrategyDefinition.match(
-        selector={"source": "A", "category": ["1.A", "1.B"]}, fill_ts=da
-    )
-    assert not primap2.csg.StrategyDefinition.match(
-        selector={"source": "A", "category": "1"}, fill_ts=da
-    )
-    assert not primap2.csg.StrategyDefinition.match(
-        selector={"source": "A", "category": ["1", "2"]}, fill_ts=da
-    )
-
-
-def test_selector_match_single_dim():
-    assert primap2.csg.StrategyDefinition.match_single_dim(
-        selector={"source": "A"}, dim="source", value="A"
-    )
-    assert not primap2.csg.StrategyDefinition.match_single_dim(
-        selector={"source": "B"}, dim="source", value="A"
-    )
-    assert primap2.csg.StrategyDefinition.match_single_dim(
-        selector={"source": ["A", "B"], "category": "1.A"}, dim="source", value="A"
-    )
-    assert primap2.csg.StrategyDefinition.match_single_dim(
-        selector={"source": "A", "category": "1"}, dim="other", value="any"
-    )
-
-
-def test_strategy_definition():
-    da = get_single_ts(coords={"source": "A", "category": "1.A"})
-
-    assert (
-        primap2.csg.StrategyDefinition(
-            [({"source": "A", "category": "1"}, 1), ({"source": "A"}, 2)]
-        ).find_strategy(da)
-        == 2
-    )
-    assert (
-        primap2.csg.StrategyDefinition(
-            [
-                ({"source": "A", "category": "1"}, 1),
-                ({"source": "A", "category": "1.A"}, 2),
-            ]
-        ).find_strategy(da)
-        == 2
-    )
-    with pytest.raises(KeyError):
-        primap2.csg.StrategyDefinition(
-            [
-                ({"source": "A", "category": "1"}, 1),
-                ({"source": "A", "category": "1.B"}, 2),
-                ({"source": "B", "category": "1.B"}, 3),
-            ]
-        ).find_strategy(da)
-
-
-def test_strategy_definition_limit():
-    assert primap2.csg.StrategyDefinition(
-        [({"entity": "A", "source": "S"}, 1), ({"source": "T"}, 2)]
-    ).limit("entity", "A").strategies == [({"source": "S"}, 1), ({"source": "T"}, 2)]
-    assert primap2.csg.StrategyDefinition(
-        [({"entity": "A", "source": "S"}, 1), ({"source": "T"}, 2)]
-    ).limit("entity", "B").strategies == [({"source": "T"}, 2)]
-
-
-def test_priority_limit():
-    pd = primap2.csg.PriorityDefinition(
-        priority_dimensions=["a", "b"],
-        priorities=[
-            {"a": "1", "b": "2", "c": "3", "d": ["4", "5"]},
-            {"a": "2", "b": "3"},
-        ],
-    )
-    assert pd.limit("e", "3") == pd
-    assert pd.limit("c", "3").priorities == [
-        {"a": "1", "b": "2", "d": ["4", "5"]},
-        {"a": "2", "b": "3"},
-    ]
-    assert pd.limit("c", "4").priorities == [{"a": "2", "b": "3"}]
-    assert pd.limit("d", "4").priorities == [
-        {"a": "1", "b": "2", "c": "3"},
-        {"a": "2", "b": "3"},
-    ]
-    assert pd.limit("d", "5") == pd.limit("d", "4")
-    assert pd.limit("d", "6").priorities == [{"a": "2", "b": "3"}]
-
-
-def test_priority_check():
-    primap2.csg.PriorityDefinition(
-        priority_dimensions=["a", "b"],
-        priorities=[
-            {"a": "1", "b": "2", "c": "3", "d": ["4", "5"]},
-            {"a": "2", "b": "3"},
-        ],
-    ).check_dimensions()
-
-    with pytest.raises(ValueError):
-        primap2.csg.PriorityDefinition(
-            priority_dimensions=["a", "b"],
-            priorities=[
-                {"a": "1", "b": "2", "c": "3", "d": ["4", "5"]},
-                {"a": "2"},
-            ],
-        ).check_dimensions()
-
-    with pytest.raises(ValueError):
-        primap2.csg.PriorityDefinition(
-            priority_dimensions=["a", "b"],
-            priorities=[
-                {"a": "1", "b": "2", "c": "3", "d": ["4", "5"]},
-                {"a": "2", "b": ["2", "3"]},
-            ],
-        ).check_dimensions()
+from primap2.tests.csg.utils import get_single_ts
 
 
 def test_compose_simple():
